fixed target detector
onmousedown bug (not being able to place down tower due to tower collider) - I added a layer so that it only detect the collider of tiles.

not targeting enemy already inside the collider - I made List of all the enemies detected inside the collider and if the current target died or exit the collider it will get a new target from the list.